+++ src/modules/slimer-sdk/bootstrap.js
@@ -1,0 +1,9 @@
+    
+    // load the coffee-script module so it can register .coffee extensions
+    require('coffee-script/coffee-script');
+    
+    var fs = require('fs');
+       
+    fs.readFileSync = function(path, encoding) {
+        return fs.read(path);
+    }